backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
chromadb
google-generativeai
python-dotenv
langchain-text-splitters
Pillow
pyjwt
cryptography
